Displaying instance details in reshape complete message
